[#156853178] Fix: Restructured routes
