+++ server/routes/v1/meals.js
@@ -1,9 +1,8 @@
 import express from 'express';
-import MealController from '../controllers/mealController';
+import MealController from '../../controllers/mealController';
 
 const router = express.Router();
 
-// These routes are to be prepended with "/meals"
 router.get('/', MealController.getMeals);
 router.post('/', MealController.createMeal);
 router.get('/:mealId', MealController.getMeal);
