+++ server/routes/v1/menu.js
@@ -1,10 +1,7 @@
 import express from 'express';
-import MenuController from '../controllers/menuController';
+import MenuController from '../../controllers/menuController';
 
 const router = express.Router();
-
-// These routes are to be prepended with "/menus"
-// router.post('/', MenuController.createTodaysMenu);
 
 router.get('/', MenuController.getTodaysMenu);
 router.post('/', MenuController.createMenu);
